requirements.txt
python-dotenv
pydantic
pymupdf
python-docx
rich
pytest
streamlit
